edit battery driver

+++ tobas_hardware/tobas_t1/tobas_t1_core/examples/battery.cpp
@@ -8,7 +8,7 @@ using namespace std;
 int main()
 {
   t1::Battery battery;
-  double voltage, current;
+  float voltage, current;
 
   if (!battery.initialize()) {
     cerr << "Failed to initialize ADC." << endl;
